Move concrete classes to 'impl' subpackages
