+++ src/main/java/edu/upc/etsetb/arqsoft/miniexceljc/factories/impl/DefaultFactory.java
@@ -9,7 +9,12 @@ import edu.upc.etsetb.arqsoft.miniexceljc.factories.SpreadsheetFactory;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.CoordinateSpec;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.NumericalValue;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions.*;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions.impl.MaxFunction;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions.impl.MinFunction;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions.impl.PromedioFunction;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions.impl.SumaFunction;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.operands.*;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.operands.impl.*;
 import edu.upc.etsetb.arqsoft.miniexceljc.postfix.*;
 import edu.upc.etsetb.arqsoft.miniexceljc.postfix.impl.PostFixGeneratorImpl;
 import edu.upc.etsetb.arqsoft.miniexceljc.postfix.impl.SyntaxCheckerImpl;
