+++ src/main/java/edu/upc/etsetb/arqsoft/miniexceljc/model/domain/functions/impl/MinFunction.java
@@ -1,8 +1,9 @@
-package edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions;
+package edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions.impl;
 
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.NumericalValue;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Spreadsheet;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Value;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions.Function;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.operands.Operand;
 
 import java.util.ArrayList;
@@ -29,7 +30,7 @@ public class MinFunction implements Function {
     public List<Value> getValue(Spreadsheet spreadsheet) {
         double min = Double.POSITIVE_INFINITY;
         for (Operand operand: this.elements) {
-            for (Value value: operand.getValue()) {
+            for (Value value: operand.getValue(spreadsheet)) {
                 double candidate = ((NumericalValue) value).getNumber();
                 if (min > candidate)
                     min = candidate;
