+++ src/main/java/edu/upc/etsetb/arqsoft/miniexceljc/model/domain/functions/impl/PromedioFunction.java
@@ -1,8 +1,9 @@
-package edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions;
+package edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions.impl;
 
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.NumericalValue;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Spreadsheet;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Value;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions.Function;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.operands.Operand;
 
 import java.util.ArrayList;
@@ -30,7 +31,7 @@ public class PromedioFunction implements Function {
         double sum = 0;
         int numElements = 0;
         for (Operand operand: this.elements) {
-            for (Value value: operand.getValue()) {
+            for (Value value: operand.getValue(spreadsheet)) {
                 sum += ((NumericalValue) value).getNumber();
                 numElements++;
             }
