+++ src/main/java/edu/upc/etsetb/arqsoft/miniexceljc/model/domain/functions/impl/SumaFunction.java
@@ -1,8 +1,9 @@
-package edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions;
+package edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions.impl;
 
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.NumericalValue;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Spreadsheet;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Value;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.functions.Function;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.operands.Operand;
 
 import java.util.ArrayList;
@@ -29,7 +30,7 @@ public class SumaFunction implements Function {
     public List<Value> getValue(Spreadsheet spreadsheet) {
         double sum = 0;
         for (Operand operand: this.elements) {
-            for (Value value: operand.getValue()) {
+            for (Value value: operand.getValue(spreadsheet)) {
                 sum += ((NumericalValue) value).getNumber();
             }
         }
