+++ src/main/java/edu/upc/etsetb/arqsoft/miniexceljc/model/domain/operands/impl/DivOperator.java
@@ -1,9 +1,11 @@
-package edu.upc.etsetb.arqsoft.miniexceljc.model.domain.operands;
+package edu.upc.etsetb.arqsoft.miniexceljc.model.domain.operands.impl;
 
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.NumericalValue;
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.Value;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.operands.Operand;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.operands.Operator;
 
-public class MultOperator implements Operator {
+public class DivOperator implements Operator {
     @Override
     public boolean isAdd() {
         return false;
@@ -16,18 +18,18 @@ public class MultOperator implements Operator {
 
     @Override
     public boolean isMult() {
-        return true;
+        return false;
     }
 
     @Override
     public boolean isDiv() {
-        return false;
+        return true;
     }
 
     @Override
     public Value operate(Operand op1, Operand op2) {
         NumericalValue value1 = (NumericalValue) op1.getValue();
         NumericalValue value2 = (NumericalValue) op2.getValue();
-        return new NumericalValue(value1.getNumber() * value2.getNumber());
+        return new NumericalValue(value1.getNumber() / value2.getNumber());
     }
 }
