+++ src/main/java/edu/upc/etsetb/arqsoft/miniexceljc/model/domain/operands/impl/RangeOperand.java
@@ -1,6 +1,7 @@
-package edu.upc.etsetb.arqsoft.miniexceljc.model.domain.operands;
+package edu.upc.etsetb.arqsoft.miniexceljc.model.domain.operands.impl;
 
 import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.*;
+import edu.upc.etsetb.arqsoft.miniexceljc.model.domain.operands.Operand;
 import edu.upc.etsetb.arqsoft.miniexceljc.util.AlphabeticRadixConverter;
 
 import java.util.ArrayList;
